+++ src/gui/dialogs/screenshot_notification.cpp
@@ -54,7 +54,7 @@ void screenshot_notification::pre_show(window& window)
 
 	text_box& path_box = find_widget<text_box>(&window, "path", false);
 	path_box.set_value(filesystem::base_name(path_));
-	path_box.set_selection(0, path_box.text().find_last_of('.')); // TODO: do this cleaner!
+	path_box.set_selection(0, static_cast<int>(path_box.text().find_last_of('.'))); // TODO: do this cleaner!
 	window.keyboard_capture(&path_box);
 	connect_signal_pre_key_press(path_box, std::bind(&screenshot_notification::keypress_callback, this,
 		std::placeholders::_3, std::placeholders::_5));
